+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.0.1 [unreleased]
 
 - Added `scope` option for partial reindex
+- Added support for Japanese, Polish, and Ukranian
 
 ## 3.0.0